energy_agents/pyproject.toml
[project]
name = "energy-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==3.1.7",
    "email-validator>=2.2.0",
    "fastapi>=0.116.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
    "fastapi>=0.116.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
    "requests",
    "fastapi>=0.116.1",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.19",
    "langgraph>=0.6.5"
]
